+++ RELEASENOTES.md
@@ -10,6 +10,7 @@ AL-Go for GitHub allows you to build and test using insider builds without any e
 
 ### Issues
 - Issue 730 Support for external rulesets.
+- Issue 739 Workflow specific KeyVault settings doesn't work for localDevEnv
 
 ### New Settings
 - `enableExternalRulesets`: set this setting to true if you want to allow AL-Go to automatically download external references in rulesets.